python-service/requirements.txt
runware==0.4.0
flask==3.0.3
flask-cors==5.0.0
python-dotenv==1.0.1
requests==2.32.3
